+++ pygdtf/protocols.py
@@ -1,0 +1,107 @@
+from typing import List, Optional, Union, Dict
+from xml.etree import ElementTree
+from xml.etree.ElementTree import Element
+from .base_node import BaseNode
+from .value import *  # type: ignore
+
+
+class Rdm(BaseNode):
+    def __init__(
+        self,
+        manufacturer_id: int = 0,
+        device_model_id: int = 0,
+        software_versions: Optional[List["SoftwareVersionId"]] = None,
+        *args,
+        **kwargs,
+    ):
+        self.manufacturer_id = manufacturer_id
+        self.device_model_id = device_model_id
+        if software_versions is not None:
+            self.software_versions = software_versions
+        super().__init__(*args, **kwargs)
+
+    def _read_xml(self, xml_node: "Element", xml_parent: Optional["Element"] = None):
+        self.manufacturer_id = int(xml_node.attrib.get("ManufacturerID", "0"), 16)
+        self.device_model_id = int(xml_node.attrib.get("DeviceModelID", "0"), 16)
+        self.software_versions = [
+            SoftwareVersionId(xml_node=i) for i in xml_node.findall("SoftwareVersionID")
+        ]
+
+    def __str__(self):
+        return (
+            f"{self.manufacturer_id} ({self.device_model_id}) {self.software_versions}"
+        )
+
+
+class SoftwareVersionId(BaseNode):
+    def __init__(
+        self,
+        value: Optional[str] = None,
+        dmx_personalities: Optional[List["DmxPersonality"]] = None,
+        *args,
+        **kwargs,
+    ):
+        self.value = value
+        if dmx_personalities is not None:
+            self.dmx_personalities = dmx_personalities
+        super().__init__(*args, **kwargs)
+
+    def _read_xml(self, xml_node: "Element", xml_parent: Optional["Element"] = None):
+        self.value = xml_node.attrib.get("Value")
+        self.dmx_personalities = [
+            DmxPersonality(xml_node=i) for i in xml_node.findall("DMXPersonality")
+        ]
+
+    def __str__(self):
+        return f"{self.value} {self.dmx_personalities}"
+
+
+class DmxPersonality(BaseNode):
+    def __init__(
+        self,
+        dmx_mode: Optional[str] = None,
+        value: Optional[str] = None,
+        *args,
+        **kwargs,
+    ):
+        self.dmx_mode = dmx_mode
+        self.value = value
+        super().__init__(*args, **kwargs)
+
+    def _read_xml(self, xml_node: "Element", xml_parent: Optional["Element"] = None):
+        self.dmx_mode = xml_node.attrib.get("DMXMode")
+        self.value = xml_node.attrib.get("Value")
+
+    def __str__(self):
+        return f"{self.dmx_mode} ({self.value})"
+
+
+class Map(BaseNode):
+    def __init__(
+        self,
+        key: int = 0,
+        value: int = 0,
+        *args,
+        **kwargs,
+    ):
+        self.key = key
+        self.value = value
+        super().__init__(*args, **kwargs)
+
+    def _read_xml(self, xml_node: "Element", xml_parent: Optional["Element"] = None):
+        self.key = int(xml_node.attrib.get("Key", 0))
+        self.value = int(xml_node.attrib.get("Value", 0))
+
+    def __str__(self):
+        return f"{self.key} {self.value}"
+
+
+class ArtNet(BaseNode):
+    def __init__(
+        self,
+        maps: Optional[List["Map"]] = None,
+        *args,
+        **kwargs,
+    ):
+        if maps is not None:
+            self.maps = maps
